Feature: Rename YoutubeProcessor to YoutubeExtractor
